Kernel mapped to higher-half virtual address.
Mapped the kernel to virtual address 0xC0028000. Kernel is still loaded
at the same 0x28000 physical location.

Two page directories are filled to give the following mapping.
1. Identity mapped: 0x0000 to 0x3FFFFF.
2. Kernel mapping: 0xC0000000 to 0x00000.

There is one page table, so 4 MB can be accessed. The highest address that can be accessed is
1. 0x3FFFFF (identity mapped to 0x3FFFFF physical address).or
2. 0xC03FFFFF (mapped to 0x3FFFFF physical address).

Work is still pending. This is more a POC.