Bug Fix: k_assert produced Page Fault exception.
This is caused, because the recent code style change, modified the
call to k_panic from the k_assert macro like so.
k_assert macro like so.

    k_panic ("Assertion failed:%s.\r\n % s", #t,e)

That extra space in '% s', for some reason caused the page fault. The
exact root cause is not known yet. But because, this is clearly a miss
from the past code style change, I fixed it here.