Bug fix:[bug-address-colision-A] Address collision
Bug:
'boot1' calculated the kernel size, without taking the .bss section into
account. This means, subsequent operations that depend on this size,
are done incorrectly, thus override memory allocated for the .bss
section.

Fix:
Moved the .bss section into the .data section. The kernel size is now
equal the kernel flat binary file size. This solution is simple and
works.

Also in this commit:
* Along with the .bss section, the .data section also contains
  COMMON objects. These are objects that are declared, but not
  initialized. Instead of the .bss section, they are taken as
  a COMMON symbol.
* .data, .rodata, .bss are now aligned to 8 byte boundaries.