Fix unittest: conflicting type size_t
Redefining the builtin type 'size_t' caused conflict with the stddef.h
which also defines the same type. This previously did not occur, because
stddef.h files was not used in MeghaOS. However, it is used in the
unittesting framework. To remedy this conflict, I included the
freestanding header, stddef.h in MeghaOS as well.

As a rule, we must not try to redefine any stdlib types, or must be
guarded to not redefine when building for Unittests.