Include header files when building unittests
Unittests are build for the host architecture, which may differ from the
target architecture of the OS. The unittests however have the same
dependencies as the OS, so the same header files must get included.

Previously, header files were included based only on the target
architecture, but now when building unittests, the same header files
will be included, no matter what the host architecture is.