Unittest fix: Handles overflow conditions.
If the output is truncated due to 'size' limit, then the return value is
the number of characters (excluding the terminating null byte) which
would have been written to the final string if enough space had been
available.

Thus, a return value of size or more means that the output was truncated.

This above behaviour is same as the snprintf implementation provided in
gcc libc.