Support for architecture specific mock & unittests
The change allows for easy addition of both architecture dependent and
independent tests. The build framework does this by introducing an
interface for specifying test dependencies and keeps most parts common
for both these types of tests.

1. Provides a structured way to support architecture dependent mock
   files in an architecture independent test.
2. Provides a structured way to build architecture dependent tests,
   while keeping most of the building and linking parts common.

Also includes one fix:
BugFix: The parent ./build.sh did not stop, if there is an error in
        unittests/build.sh.