Updated to latest meghatest [commit 424336b50]
Mock function declaration macros are removed.

The functions would already be declared in the header files of the
project. So it makes no sense to declare again for fake functions.
This change thus forces the inclusion of the project header files
to fake C files and thus the original declaration, will be taken
and used.