Dynamic return type feature is removed.
Return types are now what is given in the declaration/definition. It is
no longer an array of the given type.
This is done to simplify the default behaviour, as the array created
complication when return type is a pointer type (required to always set
before use. Otherwise, the line 'return ret[0]' will seg fault.)