BugFix: 'seg:off' calculation in SEG_MEM_LOC fixed.
Because of the overlapping nature of segments, it is not possible to
calculate the correct (as required) segment and offset values from
a Physical address. However the reverse is possible.

SEG_MEM_LOC replaces the previous MEM_LOC macro. It takes a
32 bit number (packed as 16 bit segment and 16 bit offset) and
calculates the Memory address, not the other way around like the
previous MEM_LOC macro.

New MEM_LOC just defines a new non segmented memory location.

Also in this commit:
* FILE_BUFF_MEM used instead of explicit calculation in boot1.s
* Updated BOOT1_BUILD.