BugFix: PMM alloc/deallocating past available ram
When checking for valid page and address ranges in
kpmm_alloc(At), kpmm_free etc, the MAX_ADDRESSABLE_BYTE_COUNT,
MAX_ADDRESSABLE_PAGE_COUNT was used. However this is the absolute
maximum, if system can have RAM than this amount.

Now the function kpmm_getActualAddressablePageCount and
kpmm_getActualAddressableRamSize takes the place where originally the
above macros were used.

As this was a logical error, unittests also were not able to catch this.

Also in this commit:
* Old unittests updated.
* New unittests added for the kpmm_getActualAddressableRamSize function.