Fix: Arch of headers files must be specified
Previously the `include/x86` was added to the search path, this was
inline with the previous 'single header file' approach, but as we are
now including specific header files, it makes more sense for includes to
reflect complete relative path of header files.

Now only `include` folder is in the search path and all header files
must be specified relative to it.