Renamed include/errno.h -> include/kerror.h
Renamed due to name conflict with library errno.h header. Also renamed
macros EXIT_SUCCESS and EXIT_FAILURE also due to conflict with macros of
same name in stdlib.h.

Both the library headers cannot be used in freestanding environment,
so the conflicts only happened in the unittests not in the kernel.