Unit test specific k_assert and k_panic
k_panic: When building for unitests, displays a shorter text without
filename and line number etc. This was done because the default panic
handler (in an unittest) will show this text along with the rest of the
unittest text.

k_assert: If there was an assertion failure, then its assertion
condition is not checked and calls panic immediately.

This handles the situation where there are multiple k_assert calls
in a function (which is being tested). In normal operation the first
k_assert call would have halted the system, and the second assert will
not be executed, not so in unittests, this causes problem for example
in the below code:

```
k_assert(bitmap != NULL, "Pointer to bitmap is NULL.");
k_assert(bitmap->allow != NULL, "Pointer to allow function is NULL.");
```
When `bitmap = NULL`, executing the second assert will segfault. This is
now prevented.

Also in this commit:
* k_panic_ndu panic hander can now be shared among unittests by
  including `unittests/common.c` file. Helps in displaying consistent
  output from unittests.
* Minor change in formatting of code.