Fix: Printing unchecked asserting in Unittests
When building for unittests, previously, if a previous
asserting has failed, checking of assertion condition is skipped and
panic was called. This was a mistake as now panic prints assertion
failure message when the assertion was skipped.

Now, if a previous asserting has failed, k_assert skips everything. So
after the 1st failure assert, no other assert gets hit.

Also in this commit:
* When building for unittests, k_panic now prints function name and line
  number also.