Explicit panic handling not required in functions
Previously, when built for unittests asserts/panics (in the unit under
test) would continue executing and required explicit checks to return
from the function if `panic_invoked` is true.

This was a very bad design - Kernel functions needed be aware about
unittest framework and setup and must implement code for unittests to
work properly.

Now, the panic macro can return on its own function if hit.

ISSUE:
* Must be replicated for every host where unittests is to run.
* Hard-coded x86 assembly in an arch independent header.

Also in the commit:
* Minor changes to make unittests build & pass when not building for
  DEBUG.