Physical address 0 is now a valid address
I think treating physical address 0 as an invalid NULL address made
little sense. This prevented the use of the whole 1st page and calling
PMM interfaces became difficult for the caller should check for invalid
addresses and adjusted accordingly.

The ease of debugging argument is valid, but it does not make sense how
it makes the use of the interfaces more difficult. So I think we should
setup the PAB when so required for debugging, but there is no need to
implement PMM around this idea.