PMM: PAB storing 2 bits/state (incomplete)
Each page is now represented in the PAB using 2 bits. The state of each
page can now be FREE, USED, RESERVED and INVALID. The PAB size remains
4096 bytes, but because each byte now stores state of 4 pages and not 8
as before, the maximum memory which can be accessed now is 64 MiB and
not 128 MiB as before.

Areas which are not present in the BIOS memory map is treated as
reserved.

NOTE: pmm_test is currently failing.