Fix: Separate stack needed for Kernel processes
Stack will get corrupted when there are multiple kernel processes all
using the same Stack.

The difference is when doing system call. In case of a user process
there will be one process and a Kernel stack. Kernel stack will be used
when in ring 0 (Kernel stack is common among all processes).

In case of a kernel process, there is just one process stack (no ring 0
stack is required as the whole application runs in ring 0) and the same
is used when inside a system call.

Also note that when we transfer to a kernel process, just the stack
pointer changes, all other segment registers need not change (because
we need to be in Kernel mode already to 'switch' to a process).