Thread type process implemented
Threads are processes which run in the address space of its parent. New
binary is not loaded but runs a function from the parent's binary.

Threads are processes and therefore have separate stacks. They can be
run in Kernel mode or in user mode just like regular processes. A user
mode process can have a Kernel mode thread and vice versa - Freedom!