Fix: Switching to a thread from a thread caused fault
This is caused because we attempted to map the stack of the new process
in the same location where the stack of its parent is mapped.

This fix also addresses another stack corruption when a process (thread
or otherwise) creates a new non-thread process. The context was switched
too early causing all the local state getting lost before the final jump
to the process can occur.