Fix: Unable to switch back to a Kernel process.
At the time of switching to another process the current stack states and
the return location is stored in the ProcessInfo of the current process.
For Kernel processes, the state states are not part of the interrupt
frame so these values could not be saved which resulted in garbage stack
state when returning to a Kernel process.

I now create a "System call frame" from the interrupt frame and to keep
it consistent, it always has the SS:ESP and EBP of the calling process
(Kernel process or otherwise). This solves the problem with the Kernel
process stack state preservation.