Unused ENTRY command removed from linker script
ENTRY command in the linker script does not ensure that the entry
point function is the first instruction to be executed in a flat
binary. It is the responsibility of the ELF loader of the OS to
jump to the entry point as found in the ELF header. As I do not
have ELF loader there is no use of the ENTRY command.

With function reordering option disabled the functions will appear in
the order they appear in the source. The first instruction in the source
becomes the first instruction executed.