cmake: Generate Listing files after linking
It is done at the time of linking rather than compiling because the
format of output file from compile_lib depends on the program and flags
set to produce the output. For example when compile_lib is used to
compile assembly file using NASM With `-f bin` flag, the output file
format will be binary. This makes it difficult to use `objdump` as it
does not accept a flat binary files.

This issue is not there at the time of linking, in my case it will
always produce an ELF file.

Also includes:
* Minor changes to top level CMakeLists.txt
* Cross compiled objdump is used to generate the listing files.