Now targeting i686 not i386
Along with this the optimization level for the Kernel and user programs
was changed from '-O1' to '-Os'. This enables 'O2' level of optimization
without those which increase binary size.

Related to the above change, '-fno-inline-small-functions' is required
for mpdemo and proc1 programs otherwise 'syscall' function gets inlined
which causes problems.