Stores process/Kernel PD within VMemoryManager object
VMM works in the context of a page directory. It stores a shadow copy of
that the PD should look like. So it is makes sense for VMM to own the PD
it works on. Previously only a pointer was stored which means that
someone owns the PD and VMM is just borrowing it. Now PDs are stored in
one place - inside VMemoryManager struct.