New kboot_getGraphicsModeInfo function
Note: Until now every boot structures were opaque and getter function
were used access the structure fields, but the new bootloader structure
for Graphics 'GraphisModeInfo' is not opaque. Here is the reason and
some background.

Data passed from Bootloader is grouped in structures - BootFileItem,
BootMemoryMapItem etc. These structures have fields, and there is one
function for each of the fields in these structures. The main goal for
this was abstraction; to separate out the bootloader structures from the
users of the data. If someday we start using a new bootloader, and these
structures change, only the implementation of the getters would change
and users of them would require no change.

However what if instead of getter for each fields, we group them into
structures and have getter for each of the structure. In this case also
the same abstraction can be achieved but with far less number of
getters.

This is what I have tried to do with the new GraphisModeInfo structure.

Also in this commit:
* Fix: Handles change xResolution != BytesPerScanLine case, resulting
  in incomplete fill the frame buffer