Fix: EFLAGS was not setup properly for new processes
Previously EFLAGS was set to 0, which is itself a wrong value, since
EFLAGS.Bit1 must always be set to be valid.

It now also sets the EFLAGS.IF flag to keep interrupts enabled while the
process/thread is running. This one change will keep the interrupts
enabled throughout the normal operation of the OS as the register gets
saved and restored through yield.

Also includes:
* New x86_READ_EFLAGS macro to read elfags register.