More logical driver source path hierarchy
Changed hierarchy from MARCH/ARCH to ARCH/MARCH.

This was done because most of the arch specific files need not be
machine specific. For us Machine architecture is a subset of Processor
architecture not the other way around. I think this makes more sense.
Also if we want to add a machine dependent source, we can create a MARCH
folder within the ARCH folder already present.

Also includes:
* Removed some hard-coded 'x86' with '$ARCH' in cmake files.