Macro for PAUSE X86 instruction added
Added in Pentium4, it helps CPU performance when exiting loop. Read more
in the Intel Instruction set Volume 2.

This instruction is backwards compatible so does not effect when run in
earlier CPUs. When run in Pentium4 or later would provide loop
optimisations. At least that what I think would happen.