Fix: Random hanging of gpu0 demo
Root cause is the timer interrupt taking a longer time to complete which
ends in sending EOI just as PIC is in the middle of generating the next
timer interrupt. On receiving a EOI PIC clears the interrupt before it
can complete its handshaking with the CPU, this causes it to raise the
default IRQ7. The hanging occurred because there was no handler for IRQ7
causing a GP fault. Nothing was seen on the screen because k_panic at
this point does not work in graphical mode.

See the /docs/notes/BUGS.md for some more detail.