Build definitions don't apply for unittests
It was wrong to assume build definitions for unittests. Whether to build
with DEBUG or GRAPHICS_MODE_ENABLED etc would depend on the test and
should be assumed at the global level.

cmake function `compile_lib` & `test` now accepts compiler DEFINITIONS
so that unittests can pass in the definitions it requires.