cm: New header file for private CM types/func declaration
This new file would hold types/function/macro definitions which are
private to the CM library and should not be exposed to the applications.

CM library is spread out in multiple files and there are cases when a
function/type which is defined in one file need to be exposed to another
file such that its not exposed to the applications.

Also includes:
* `CM_RETURN_ERROR` macro definition and definition of the `cm_err_num`
  global variable were moved to `kcmlib.h` since they are a private
  items.