Disable SSE, FPU CPU features in Zig builds
This makes the compiler not emit code for MMX, SSE and FPU features. The
'soft_float' is enabled because otherwise for the compiler throws an
error. I think LLVM requires some floating point support and soft_float
satisfies that.

The error I get is the following:
LLVM ERROR: Do not know how to soften this operator's operand!

Note:
* soft_float works but produces very large binaries as you can imagine
  it would, so its best to not use floating point too much or at all
  right now.